🐛 fix(runner): Make all test hierarchy levels navigable
Previously, you could only click on an it block in the test runner to
jump to the code. Clicking on a describe suite didn't do anything.

The test location provider is now smarter. It first tries to find a
specific test. If it can't find one, it assumes you clicked a suite and
looks for that instead.