⚡️♻️ Refactor: Switch conf file discovery to ProjectFileIndex
The previous method for finding the wdio.conf.js file used a recursive file search. This approach was inefficient on large projects and didn't respect project settings like "excluded" folders, which could lead to incorrect behavior. Now we use the ProjectFileIndex which is near instantaneous and respects the users' excluded files.